+++ deny.toml
@@ -1,20 +1,18 @@
+[graph]
+all-features = true
+
 [advisories]
+version = 2
 db-path = "~/.cargo/advisory-db"
 db-urls = ["https://github.com/rustsec/advisory-db"]
-vulnerability = "deny"
-unmaintained = "deny"
 yanked = "deny"
-notice = "deny"
 ignore = [
 ]
 
 [licenses]
-unlicensed = "deny"
 allow = [
     "Apache-2.0",
     "MIT",
     "Unicode-DFS-2016",
 ]
-copyleft = "deny"
-default = "deny"
 confidence-threshold = 1.0
