+++ motd/Cargo.toml
@@ -14,17 +14,22 @@ tool directly.
 license = "Apache-2.0"
 keywords = ["motd", "ssh", "terminal", "shell"]
 
+[features]
+default = ["socall"]
+socall = ["dep:lazy_static", "dep:libc", "dep:pam-sys", "dep:tempfile",
+          "dep:serde_json", "dep:dlopen2", "dep:walkdir"]
+
 [dependencies]
-lazy_static = "1"
-libc = "0.2"
 log = "0.4"
-pam-sys = "0.5"
-tempfile = "3"
-serde = "1"
 serde_derive = "1"
-serde_json = "1"
-dlopen2 = "0.7.0"
-walkdir = "2.5.0"
+serde = "1"
+lazy_static = { version = "1", optional = true }
+libc = { version = "0.2", optional = true }
+pam-sys = { version = "0.5", optional = true }
+tempfile = { version = "3", optional = true }
+serde_json = { version = "1", optional = true }
+dlopen2 = { version = "0.7.0", optional = true }
+walkdir = { version = "2.5.0", optional = true }
 
 [build-dependencies]
 which = "6"
